Change 一度木をすべて伐採 キャラクターの見た目を変更するために、新たなキャラクター追加

+++ Assets/MyGameAssets/Scripts/QuestionScript/Question.cs
@@ -17,8 +17,8 @@ public class Question : MonoBehaviour
     [SerializeField]
     private TextMeshPro answerText;
 
-    private int firstNumber;
-    private int secondNumber;
+    private float firstNumber;
+    private float secondNumber;
 
     private float answer;
     private float playerAnswer;
